+++ Mods ChoGGi/Rover Icons Map Overview/metadata.lua
@@ -1,0 +1,16 @@
+return PlaceObj("ModDef", {
+	"title", "Rover Icons Map Overview v0.1",
+	"version", 1,
+	"saved", 1543924800,
+	"image", "Preview.png",
+	"id", "ChoGGi_RoverIconsMapOverview",
+	"steam_id", "1582759681",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[Makes rovers larger on zoom (same as res icons)
+
+Requested by oIMADMANIo.]],
+})